pyproject.toml
[project]
name = "football-pool"
version = "0.1.0"
description = ""
authors = [
    {name = "Devilfish281",email = "Dobley@csu.fullerton.edu"}
]
readme = "README.md"
packages = [{ include = "football_pool_picker", from = "src" }]


requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.6.6,<0.7.0)",
    "openai (>=1.102.0,<2.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "langchain-core (>=0.3.75,<0.4.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "tiktoken (>=0.11.0,<0.12.0)"
]

[tool.poetry]
packages = [{include = "football_pool", from = "src"}]

[tool.poetry.scripts]
football-pool-picker = "ui.app:main"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
